Finished Level 1
Level 1 is done... The color difference between a goal that had been hit
and one that hadn't was so subtle that I decided to make the difference
a little bigger. Don't know how many castles we need to build, so I am
holding off on the last few steps till I get guidance from the
professor.