According to the book I'm done
I feel like it is missing something, I also don't like the walls being
white because they are hard to see in contrast to the clouds. I feel
like the clouds are in the wrong layer or something, so I may recolor
and fix some things to tweak this to how I want it.